Initial commit for AE RPSL demo
Includes some limited basic functionality in the index page with the
player always choosing "rock".  AI randomly selects a "weapon" and then
the "fight" method is called.  The method will determine if there is a
draw first, if not a winner is determined.  The method returns an
associative array to the page.